Move frontend and backend to the correct folder
Exactly what the headline says. Move this stuff to the app folder.
