+++ app/frontend/js/constants.js
