+++ app/frontend/js/snake.js
